+++ plugin-gradle/CHANGES.md
@@ -3,6 +3,8 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `3.27.0`).
 
 ## [Unreleased]
+### Added
+* Added support for JSON formatting based on [Gson](https://github.com/google/gson)
 
 ## [6.2.2] - 2022-02-09
 ### Changed