+++ plugin-gradle/CHANGES.md
@@ -4,6 +4,10 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 
 ## [Unreleased]
 
+### Changed
+
+* Use SLF4J for logging ([#1116](https://github.com/diffplug/spotless/issues/1116))
+
 ## [6.2.2] - 2022-02-09
 ### Changed
 * Bump default ktfmt `0.30` -> `0.31` ([#1118](https://github.com/diffplug/spotless/pull/1118)).